requirements.txt
django-filter
django
pillow
celery
flower
stripe
python-decouple
weasyprint
psycopg2-binary
redis